+++ app/components/omnibox/services/nested-asset-service.js
@@ -47,8 +47,6 @@ angular.module('omnibox')
       });
 
       return nestedAssets;
-
     };
-
   }
 ]);