+++ app/lizard-nxt-master-controller.js
@@ -26,6 +26,7 @@ angular.module('lizard-nxt')
    'MapService',
    'rmAllButLastAssetAndGeometry',
    'UrlService',
+   'AssetService',
 
   function ($scope,
             $rootScope,
@@ -37,7 +38,8 @@ angular.module('lizard-nxt')
             State,
             MapService,
             rmAllButLastAssetAndGeometry,
-            UrlService) {
+            UrlService,
+            AssetService) {
 
   $scope.version = version;
   $scope.tooltips = CabinetService.createTooltips();
@@ -63,6 +65,11 @@ angular.module('lizard-nxt')
   $scope.transitionToContext = function (context) {
     if (context !== State.context) {
       State.context = context;
+      if (context === 'map' && AssetService.getNonNestedAssets().length > 1) {
+        // If more than one asset became selected when user was in db context,
+        // we switch the box-type/tool to 'multi-point':
+        State.box.type = 'multi-point';
+      }
     }
 
     var overlay = angular.element('#context-transition-overlay')[0];